continue

+++ dotNet2022_8090_7731/BL/BL/DroneInParcel.cs
@@ -9,10 +9,10 @@ namespace IBL.BO
     public class DroneInParcel
     {
         public int Id{ get; set; }
-        public float BatteryStatus{ get; set; }
+        public double BatteryStatus{ get; set; }
         public Location CurrLocation { get; set; }
         
-        public DroneInParcel (int id, float batteryStatus, Location location)
+        public DroneInParcel (int id, double batteryStatus, Location location)
 	    {
             Id=id;
             BatteryStatus = batteryStatus;
